requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
backoff>=2.2.1
web3>=6.11.1
setuptools>=65.5.1
tabulate==0.9.0
pandas==2.1.4
schedule==1.2.1
matplotlib==3.8.2
seaborn==0.13.0
sqlalchemy==2.0.27
flask==3.0.2
flask-cors==4.0.0 